fix test pkg and rename to devnet
