+++ contracts/contracts_test.go
@@ -10,7 +10,7 @@ import (
 	"github.com/NethermindEth/juno/core/felt"
 	starknetgo "github.com/NethermindEth/starknet.go"
 	"github.com/NethermindEth/starknet.go/artifacts"
-	devtest "github.com/NethermindEth/starknet.go/test"
+	devtest "github.com/NethermindEth/starknet.go/devnet"
 	"github.com/NethermindEth/starknet.go/types"
 	"github.com/NethermindEth/starknet.go/utils"
 	"github.com/joho/godotenv"
