+++ devnet/main_test.go
@@ -1,0 +1,28 @@
+package test
+
+import (
+	"flag"
+	"os"
+	"testing"
+)
+
+const (
+	DevNetETHAddress = "0x49d36570d4e46f48e99674bd3fcc84644ddd6b96f7c741b1562b82f9e004dc7"
+)
+
+// testConfiguration is a type that is used to configure tests
+type testConfiguration struct {
+	base string
+}
+
+var (
+	// set the environment for the test, default: devnet
+	testEnv = "devnet"
+)
+
+// TestMain is used to trigger the tests and, in that case, check for the environment to use.
+func TestMain(m *testing.M) {
+	flag.StringVar(&testEnv, "env", "devnet", "set the test environment")
+	flag.Parse()
+	os.Exit(m.Run())
+}
